har_cascade model that is used for image classification.
har_cascade model that is used for image classification into solo and group categories. it is used keeping in mind to keep the model light weight so that it can run on the render specific constraints of 8gb of ram which others model might exceed, also it proves to be a pretty accurate model.